requirements.txt
Flask==2.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flask-jwt-extended==4.3.0
flask-cors==3.0.10
python-dotenv==0.19.0
werkzeug==2.2.0
mysqlclient==2.2.1